+++ lib/rspec/core/notifications.rb
@@ -36,14 +36,21 @@ module RSpec::Core
     class ExampleNotification
       # @private
       def self.for(example)
-        if example.execution_result.pending_fixed?
+        execution_result = example.execution_result
+
+        if execution_result.pending_fixed?
           PendingExampleFixedNotification.new(example)
-        elsif example.execution_result.status == :failed
+        elsif execution_result.example_skipped?
+          SkippedExampleNotification.new(example)
+        elsif execution_result.status == :pending
+          PendingExampleFailedAsExpectedNotification.new(example)
+        elsif execution_result.status == :failed
           FailedExampleNotification.new(example)
         else
           new(example)
         end
       end
+
       private_class_method :new
     end
 
@@ -60,31 +67,38 @@ module RSpec::Core
         @reporter = reporter
       end
 
-      # @return [Array(RSpec::Core::Example)] list of examples
+      # @return [Array<RSpec::Core::Example>] list of examples
       def examples
         @reporter.examples
       end
 
-      # @return [Array(RSpec::Core::Example)] list of failed examples
+      # @return [Array<RSpec::Core::Example>] list of failed examples
       def failed_examples
         @reporter.failed_examples
       end
 
-      # @return [Array(RSpec::Core::Example)] list of pending examples
+      # @return [Array<RSpec::Core::Example>] list of pending examples
       def pending_examples
         @reporter.pending_examples
       end
 
-      # @return [Array(Rspec::Core::Notifications::ExampleNotification]
+      # @return [Array<RSpec::Core::Notifications::ExampleNotification>]
       #         returns examples as notifications
       def notifications
         @notifications ||= format_examples(examples)
       end
 
-      # @return [Array(Rspec::Core::Notifications::FailedExampleNotification]
+      # @return [Array<RSpec::Core::Notifications::FailedExampleNotification>]
       #         returns failed examples as notifications
       def failure_notifications
         @failed_notifications ||= format_examples(failed_examples)
+      end
+
+      # @return [Array<RSpec::Core::Notifications::SkippedExampleNotification,
+      #                 RSpec::Core::Notifications::PendingExampleFailedAsExpectedNotification>]
+      #         returns pending examples as notifications
+      def pending_notifications
+        @pending_notifications ||= format_examples(pending_examples)
       end
 
       # @return [String] The list of failed examples, fully formatted in the way
@@ -102,15 +116,10 @@ module RSpec::Core
       # @return [String] The list of pending examples, fully formatted in the
       #   way that RSpec's built-in formatters emit.
       def fully_formatted_pending_examples(colorizer=::RSpec::Core::Formatters::ConsoleCodes)
-        formatted = "\nPending:\n"
+        formatted = "\nPending: (Failures listed here are expected and do not affect your suite's status)\n"
 
-        pending_examples.each do |example|
-          formatted_caller = RSpec.configuration.backtrace_formatter.backtrace_line(example.location)
-
-          formatted <<
-            "  #{colorizer.wrap(example.full_description, :pending)}\n" \
-            "    # #{colorizer.wrap(example.execution_result.pending_message, :detail)}\n" \
-            "    # #{colorizer.wrap(formatted_caller, :detail)}\n"
+        pending_notifications.each_with_index do |notification, index|
+          formatted << notification.fully_formatted(index.next, colorizer)
         end
 
         formatted
@@ -152,7 +161,7 @@ module RSpec::Core
 
       # Returns the message generated for this failure line by line.
       #
-      # @return [Array(String)] The example failure message
+      # @return [Array<String>] The example failure message
       def message_lines
         add_shared_group_lines(failure_lines, NullColorizer)
       end
@@ -160,16 +169,16 @@ module RSpec::Core
       # Returns the message generated for this failure colorized line by line.
       #
       # @param colorizer [#wrap] An object to colorize the message_lines by
-      # @return [Array(String)] The example failure message colorized
+      # @return [Array<String>] The example failure message colorized
       def colorized_message_lines(colorizer=::RSpec::Core::Formatters::ConsoleCodes)
         add_shared_group_lines(failure_lines, colorizer).map do |line|
-          colorizer.wrap line, RSpec.configuration.failure_color
+          colorizer.wrap line, message_color
         end
       end
 
       # Returns the failures formatted backtrace.
       #
-      # @return [Array(String)] the examples backtrace lines
+      # @return [Array<String>] the examples backtrace lines
       def formatted_backtrace
         backtrace_formatter.format_backtrace(exception.backtrace, example.metadata)
       end
@@ -177,7 +186,7 @@ module RSpec::Core
       # Returns the failures colorized formatted backtrace.
       #
       # @param colorizer [#wrap] An object to colorize the message_lines by
-      # @return [Array(String)] the examples colorized backtrace lines
+      # @return [Array<String>] the examples colorized backtrace lines
       def colorized_formatted_backtrace(colorizer=::RSpec::Core::Formatters::ConsoleCodes)
         formatted_backtrace.map do |backtrace_info|
           colorizer.wrap "# #{backtrace_info}", RSpec.configuration.detail_color
@@ -187,17 +196,7 @@ module RSpec::Core
       # @return [String] The failure information fully formatted in the way that
       #   RSpec's built-in formatters emit.
       def fully_formatted(failure_number, colorizer=::RSpec::Core::Formatters::ConsoleCodes)
-        formatted = "\n  #{failure_number}) #{description}\n"
-
-        colorized_message_lines(colorizer).each do |line|
-          formatted << RSpec::Support::EncodedString.new("     #{line}\n", encoding_of(formatted))
-        end
-
-        colorized_formatted_backtrace(colorizer).each do |line|
-          formatted << RSpec::Support::EncodedString.new("     #{line}\n", encoding_of(formatted))
-        end
-
-        formatted
+        "\n  #{failure_number}) #{description}\n#{formatted_message_and_backtrace(colorizer)}"
       end
 
     private
@@ -266,6 +265,24 @@ module RSpec::Core
           File.expand_path(line_path).downcase == example_path
         end
       end
+
+      def formatted_message_and_backtrace(colorizer)
+        formatted = ""
+
+        colorized_message_lines(colorizer).each do |line|
+          formatted << RSpec::Support::EncodedString.new("     #{line}\n", encoding_of(formatted))
+        end
+
+        colorized_formatted_backtrace(colorizer).each do |line|
+          formatted << RSpec::Support::EncodedString.new("     #{line}\n", encoding_of(formatted))
+        end
+
+        formatted
+      end
+
+      def message_color
+        RSpec.configuration.failure_color
+      end
     end
 
     # The `PendingExampleFixedNotification` extends `ExampleNotification` with
@@ -285,7 +302,7 @@ module RSpec::Core
 
       # Returns the message generated for this failure line by line.
       #
-      # @return [Array(String)] The example failure message
+      # @return [Array<String>] The example failure message
       def message_lines
         ["Expected pending '#{example.execution_result.pending_message}' to fail. No Error was raised."]
       end
@@ -293,9 +310,63 @@ module RSpec::Core
       # Returns the message generated for this failure colorized line by line.
       #
       # @param colorizer [#wrap] An object to colorize the message_lines by
-      # @return [Array(String)] The example failure message colorized
+      # @return [Array<String>] The example failure message colorized
       def colorized_message_lines(colorizer=::RSpec::Core::Formatters::ConsoleCodes)
         message_lines.map { |line| colorizer.wrap(line, RSpec.configuration.fixed_color) }
+      end
+    end
+
+    # @private
+    module PendingExampleNotificationMethods
+    private
+
+      def fully_formatted_header(pending_number, colorizer=::RSpec::Core::Formatters::ConsoleCodes)
+        colorizer.wrap("\n  #{pending_number}) #{example.full_description}\n", :pending) <<
+        colorizer.wrap("     # #{example.execution_result.pending_message}\n", :detail)
+      end
+    end
+
+    # The `PendingExampleFailedAsExpectedNotification` extends `FailedExampleNotification` with
+    # things useful for pending specs that fail as expected.
+    #
+    # @attr [RSpec::Core::Example] example the current example
+    # @see ExampleNotification
+    class PendingExampleFailedAsExpectedNotification < FailedExampleNotification
+      include PendingExampleNotificationMethods
+      public_class_method :new
+
+      # @return [Exception] The exception that occurred while the pending example was executed
+      def exception
+        example.execution_result.pending_exception
+      end
+
+      # @return [String] The pending detail fully formatted in the way that
+      #   RSpec's built-in formatters emit.
+      def fully_formatted(pending_number, colorizer=::RSpec::Core::Formatters::ConsoleCodes)
+        fully_formatted_header(pending_number, colorizer) << formatted_message_and_backtrace(colorizer)
+      end
+
+    private
+
+      def message_color
+        RSpec.configuration.pending_color
+      end
+    end
+
+    # The `SkippedExampleNotification` extends `ExampleNotification` with
+    # things useful for specs that are skipped.
+    #
+    # @attr [RSpec::Core::Example] example the current example
+    # @see ExampleNotification
+    class SkippedExampleNotification < ExampleNotification
+      include PendingExampleNotificationMethods
+      public_class_method :new
+
+      # @return [String] The pending detail fully formatted in the way that
+      #   RSpec's built-in formatters emit.
+      def fully_formatted(pending_number, colorizer=::RSpec::Core::Formatters::ConsoleCodes)
+        formatted_caller = RSpec.configuration.backtrace_formatter.backtrace_line(example.location)
+        fully_formatted_header(pending_number, colorizer) << colorizer.wrap("     # #{formatted_caller}\n", :detail)
       end
     end
 
@@ -343,9 +414,9 @@ module RSpec::Core
     # of the test run.
     #
     # @attr duration [Float] the time taken (in seconds) to run the suite
-    # @attr examples [Array(RSpec::Core::Example)] the examples run
-    # @attr failed_examples [Array(RSpec::Core::Example)] the failed examples
-    # @attr pending_examples [Array(RSpec::Core::Example)] the pending examples
+    # @attr examples [Array<RSpec::Core::Example>] the examples run
+    # @attr failed_examples [Array<RSpec::Core::Example>] the failed examples
+    # @attr pending_examples [Array<RSpec::Core::Example>] the pending examples
     # @attr load_time [Float] the number of seconds taken to boot RSpec
     #                         and load the spec files
     SummaryNotification = Struct.new(:duration, :examples, :failed_examples, :pending_examples, :load_time)
@@ -443,11 +514,11 @@ module RSpec::Core
     # information at the end of the test run for profiling information.
     #
     # @attr duration [Float] the time taken (in seconds) to run the suite
-    # @attr examples [Array(RSpec::Core::Example)] the examples run
+    # @attr examples [Array<RSpec::Core::Example>] the examples run
     # @attr number_of_examples [Fixnum] the number of examples to profile
     ProfileNotification = Struct.new(:duration, :examples, :number_of_examples)
     class ProfileNotification
-      # @return [Array(RSpec::Core::Example)] the slowest examples
+      # @return [Array<RSpec::Core::Example>] the slowest examples
       def slowest_examples
         @slowest_examples ||=
           examples.sort_by do |example|
@@ -472,7 +543,7 @@ module RSpec::Core
           end
       end
 
-      # @return [Array(RSpec::Core::Example)] the slowest example groups
+      # @return [Array<RSpec::Core::Example>] the slowest example groups
       def slowest_groups
         @slowest_groups ||= calculate_slowest_groups
       end